+++ CHANGELOG.md
@@ -11,6 +11,10 @@ greater clarity (#75)
 * Changed behavior of the `do_predict_proba` parameter of `environment.Environment` when `True`
     * All other behavior remains unchanged. However, instead of behaving identically to `do_predict_proba=0`,
     `do_predict_proba=True` will now use all predicted probability columns for the final predictions
+* Deprecated classes `experiments.RepeatedCVExperiment` and `experiments.StandardCVExperiment`. The 
+goals of both of these classes are accomplished by the preferred `experiments.CrossValidationExperiment`
+class. The two aforementioned deprecated classes are scheduled for removal in v2.1.0. All uses of the 
+deprecated classes should be replaced with `experiments.CrossValidationExperiment`
 
 
 <a name="2.0.0"></a>